优化系统判断,添加 vi 配色

+++ zsh/install_zsh.sh
@@ -1,5 +1,7 @@
-yum install zsh git -y
-wget https://raw.github.com/robbyrussell/oh-my-zsh/master/tools/install.sh -O - | sh
+if [ -f /etc/issue ]; then
+  yum install zsh git -y
+fi
+curl -L https://raw.github.com/robbyrussell/oh-my-zsh/master/tools/install.sh | sh
 chsh -s /bin/zsh
 home=`echo ~`
 if [ -f /etc/issue ]; then
